fix: fix swagger type

+++ src/entitys/main/artistImageVersion.entity.ts
@@ -13,7 +13,7 @@ import { Exclude } from 'class-transformer';
 @Entity({ name: 'artist_image_version' })
 export class ArtistImageVersionEntity extends BaseEntity {
   @Exclude()
-  @ApiProperty({ type: 'bigint' })
+  @ApiProperty({ type: 'int' })
   @PrimaryGeneratedColumn({ type: 'bigint' })
   id: number;
