+++ src/entitys/main/chartUpdated.entity.ts
@@ -3,7 +3,7 @@ import { BaseEntity, Entity, PrimaryColumn } from 'typeorm';
 
 @Entity({ name: 'chart_updated' })
 export class ChartUpdatedEntity extends BaseEntity {
-  @ApiProperty({ type: 'bigint' })
+  @ApiProperty({ type: 'int' })
   @PrimaryColumn({ type: 'bigint' })
   time: number;
 }
